REFACTOR: Remove all old references from Asset User Actor Network Service, plus modified PluginRoot's name.
Associated Issue: #5720
