+++ DAP/plugin/actor_network_service/fermat-dap-plugin-actor-network-service-asset-user-bitdubai/src/main/java/com/bitdubai/fermat_dap_plugin/layer/actor/network/service/asset/user/developer/bitdubai/DeveloperBitDubai.java
@@ -16,7 +16,6 @@ import com.bitdubai.fermat_api.layer.all_definition.enums.TimeFrequency;
 import com.bitdubai.fermat_api.layer.all_definition.license.PluginLicensor;
 //import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.user.developer.bitdubai.version_1.AssetUserActorNetworkServicePluginRoot;
 import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.user.developer.bitdubai.version_1.AssetUserActorNetworkServicePluginRoot;
-import com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.user.developer.bitdubai.version_1.AssetUserActorNetworkServicePluginRootNew;
 
 /**
 * The Class <code>com.bitdubai.fermat_dap_plugin.layer.actor.network.service.asset.user.developer.bitdubai.DeveloperBitDubai</code> is
@@ -37,7 +36,7 @@ public class DeveloperBitDubai extends AbstractPluginDeveloper implements Plugin
         try {
 
 //            this.registerVersion(new AssetUserActorNetworkServicePluginRoot());
-            this.registerVersion(new AssetUserActorNetworkServicePluginRootNew());
+            this.registerVersion(new AssetUserActorNetworkServicePluginRoot());
 
         } catch (CantRegisterVersionException e) {
